move the oligo-seq specific overrides directly in the oligo-seq pipeline file so that all parameters for one pipeline are in the same file
